requirements.txt
requests
google-cloud-bigquery
google-auth
google-cloud-core